requirements.txt
# === Core web ===
Flask==3.1.1
Flask-Session==0.8.0
Flask-Talisman==1.1.0
gunicorn==21.2.0
requests==2.32.5
python-dotenv==1.2.1
redis==5.2.0
psycopg[binary]>=3.2

# === Discord / Twitch ===
discord.py==2.5.2
twitchAPI==4.5.0

# === Plex ===
plex.py==0.9.0
PlexAPI==4.15.15
texttable==1.7.0

# === Web templates ===
Jinja2==3.1.6
MarkupSafe==3.0.2
itsdangerous==2.2.0
Werkzeug==3.1.3

# === Async deps (discord.py transitive) ===
aiohttp>=3.12
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
attrs==25.4.0
frozenlist==1.8.0
idna==3.11
multidict==6.7.0
propcache==0.4.1
typing_extensions==4.15.0
yarl==1.22.0
